pyproject.toml
[project]

name = "hwpw11"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
"fastapi (>=0.116.1,<0.117.0)",
"uvicorn[standard] (>=0.35.0,<0.36.0)",
"sqlalchemy (>=2.0.42,<3.0.0)",
"psycopg2-binary (>=2.9.10,<3.0.0)",
"alembic (>=1.16.4,<2.0.0)",
"asyncpg (>=0.30.0,<0.31.0)",
"pydantic[email] (>=2.11.7,<3.0.0)",
"libgravatar (>=1.0.4,<2.0.0)",
"python-jose[cryptography] (>=3.5.0,<4.0.0)",
"passlib[bcrypt] (>=1.7.4,<2.0.0)",
"python-multipart (>=0.0.20,<0.0.21)",
"pydantic-settings (>=2.10.1,<3.0.0)",
"python-dotenv (>=1.1.1,<2.0.0)",
"cloudinary (>=1.44.1,<2.0.0)",
"redis (>=6.4.0,<7.0.0)",
"sqlalchemy-utils (>=0.41.2,<0.42.0)",
"aiosmtplib (>=4.0.1,<5.0.0)",
"jinja2 (>=3.1.6,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"